@@ -20,16 +20,15 @@ import 'package:loading_animation_widget/loading_animation_widget.dart';
 class CategoryPageAppBar extends StatefulWidget {
   const CategoryPageAppBar({
     Key? key,
-    required this.contributors,
-    required this.placeList,
+    this.contributors,
+    this.placeList,
     required this.scrollController,
-    required this.listOwner,
+    this.listOwner,
   }) : super(key: key);
 
   final List<User>? contributors;
-
-  final PlaceList placeList;
-  final User listOwner;
+  final PlaceList? placeList;
+  final User? listOwner;
   final ScrollController scrollController;
 
   @override
@@ -89,112 +88,132 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Flexible(
-              flex: 4,
-              child: FittedBox(
-                child: Wrap(
-                  spacing: 16.0,
-                  crossAxisAlignment: WrapCrossAlignment.center,
-                  children: [
-                    Animate(
-                        effects: const [
-                          FadeEffect(),
-                          SlideEffect(
-                              duration: Duration(milliseconds: 400),
-                              curve: Curves.easeOutSine,
-                              begin: Offset(-1.0, 0),
-                              end: Offset(0, 0))
-                          // RotateEffect(
-                          //     curve: Curves.easeOutBack,
-                          //     duration: Duration(milliseconds: 500),
-                          //     begin: -0.25,
-                          //     end: 0.0,
-                          //     alignment: Alignment.centerLeft)
-                        ],
-                        child: Padding(
-                          padding: const EdgeInsets.only(right: 4.0, left: 4.0),
-                          child: SizedBox(
+            BlocBuilder<SavedPlacesBloc, SavedPlacesState>(
+              builder: (context, state) {
+                if (state is SavedPlacesLoading) {
+                  return Padding(
+                    padding: const EdgeInsets.only(left: 8.0),
+                    child: LoadingAnimationWidget.beat(
+                        color: FlexColor.bahamaBlueDarkSecondary, size: 20.0),
+                  );
+                }
+                if (state is SavedPlacesLoaded) {
+                  return Flexible(
+                    flex: 4,
+                    child: FittedBox(
+                      child: Wrap(
+                        spacing: 16.0,
+                        crossAxisAlignment: WrapCrossAlignment.center,
+                        children: [
+                          Padding(
+                            padding:
+                                const EdgeInsets.only(right: 4.0, left: 4.0),
+                            child: SizedBox(
                               width: 30,
                               height: 30,
-                              child: Icon(
-                                deserializeIcon(widget.placeList.icon) ??
-                                    Icons.list_alt_rounded,
-                                color: Theme.of(context).brightness ==
-                                        Brightness.dark
-                                    ? Colors.white
-                                    : Colors.grey[800],
-                                size: widget.placeList.icon
-                                        .containsValue('fontAwesomeIcons')
-                                    ? 28
-                                    : 32,
-                              )),
-                        )),
-                    Animate(
-                      effects: const [
-                        FadeEffect(),
-                        RotateEffect(
+                              child: context
+                                          .watch<SavedPlacesBloc>()
+                                          .state
+                                          .placeList
+                                          ?.icon !=
+                                      null
+                                  ? Icon(
+                                      deserializeIcon(context
+                                              .read<SavedPlacesBloc>()
+                                              .state
+                                              .placeList!
+                                              .icon) ??
+                                          Icons.list_alt_rounded,
+                                      color: Theme.of(context).brightness ==
+                                              Brightness.dark
+                                          ? Colors.white
+                                          : Colors.grey[800],
+                                      size: context
+                                              .read<SavedPlacesBloc>()
+                                              .state
+                                              .placeList!
+                                              .icon
+                                              .containsValue('fontAwesomeIcons')
+                                          ? 28
+                                          : 32,
+                                    )
+                                  : const Icon(
+                                      Icons.list_alt_rounded,
+                                      color: Colors.white,
+                                      size: 32,
+                                    ),
+                            ),
+                          ),
+                          context
+                                      .watch<SavedPlacesBloc>()
+                                      .state
+                                      .placeList
+                                      ?.name !=
+                                  null
+                              ? Text(
+                                  context
+                                      .read<SavedPlacesBloc>()
+                                      .state
+                                      .placeList!
+                                      .name,
+                                  style: Theme.of(context)
+                                      .textTheme
+                                      .titleLarge!
+                                      .copyWith(fontWeight: FontWeight.bold),
+                                )
+                              : const SizedBox.shrink(),
+                        ],
+                      ).animate().fadeIn().rotate(
                             curve: Curves.easeOutBack,
-                            duration: Duration(milliseconds: 500),
+                            duration: const Duration(milliseconds: 500),
                             begin: -0.25,
                             end: 0.0,
-                            alignment: Alignment.centerLeft)
-                      ],
-                      child: Text(
-                        widget.placeList.name,
-                        style: Theme.of(context)
-                            .textTheme
-                            .titleLarge!
-                            .copyWith(fontWeight: FontWeight.bold),
-                      ),
+                            alignment: Alignment.centerLeft,
+                          ),
                     ),
-                  ],
-                ),
-              ),
+                  );
+                } else {
+                  return const SizedBox();
+                }
+              },
             ),
             Flexible(
-              child: Animate(
-                effects: const [
-                  FadeEffect(),
-                  SlideEffect(
-                      duration: Duration(milliseconds: 700),
-                      curve: Curves.easeOutQuart,
-                      begin: Offset(0.5, 0.0),
-                      end: Offset(0.0, 0.0)),
-                ],
-                child: BlocConsumer<SavedPlacesBloc, SavedPlacesState>(
-                  listener: (context, state) {
-                    if (state is PurchasesUpdated) {
-                      Navigator.pop(context);
-                    }
-                  },
-                  builder: (context, state) {
-                    if (state is SavedPlacesLoading) {
-                      return LoadingAnimationWidget.beat(
-                          color: FlexColor.bahamaBlueDarkSecondary, size: 18.0);
-                    }
-                    if (state is SavedPlacesLoaded) {
-                      return AvatarStack(
-                        settings: RestrictedPositions(
-                            align: StackAlign.right, laying: StackLaying.first),
-                        borderWidth: 2.0,
-                        borderColor:
-                            Theme.of(context).chipTheme.backgroundColor,
-                        width: 70,
-                        height: 40,
-                        avatars: [
-                          CachedNetworkImageProvider(
-                              widget.listOwner.profilePicture!),
-                          for (User user in state.contributors)
-                            CachedNetworkImageProvider(user.profilePicture!),
-                        ],
-                      );
-                    } else {
-                      return const Center(
-                        child: Text('Error Loading Avatars!'),
-                      );
-                    }
-                  },
-                ),
+              child: BlocConsumer<SavedPlacesBloc, SavedPlacesState>(
+                listener: (context, state) {
+                  if (state is PurchasesUpdated) {
+                    Navigator.pop(context);
+                  }
+                },
+                builder: (context, state) {
+                  if (state is SavedPlacesLoading) {
+                    return const SizedBox();
+                  }
+                  if (state is SavedPlacesLoaded) {
+                    return AvatarStack(
+                      settings: RestrictedPositions(
+                          align: StackAlign.right, laying: StackLaying.first),
+                      borderWidth: 2.0,
+                      borderColor: Theme.of(context).chipTheme.backgroundColor,
+                      width: 70,
+                      height: 40,
+                      avatars: [
+                        CachedNetworkImageProvider(
+                            state.listOwner.profilePicture!),
+                        for (User user in state.contributors)
+                          CachedNetworkImageProvider(user.profilePicture!),
+                      ],
+                    ).animate().fadeIn().slideX(
+                          curve: Curves.easeOutSine,
+                          duration: const Duration(milliseconds: 400),
+                          begin: 0.5,
+                          end: 0.0,
+                        );
+                  } else {
+                    return const Center(
+                      child: Text('Error Loading Avatars!'),
+                    );
+                  }
+                },
               ),
             ),
           ],
@@ -216,7 +235,7 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
                       await showDialog(
                         context: context,
                         builder: (context) => EditListDialog(
-                          placeList: widget.placeList,
+                          placeList: widget.placeList!,
                         ),
                       ).then((value) async {
                         if (value == true) {
@@ -247,7 +266,7 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
                           context: context,
                           builder: (context) {
                             return DeleteListDialog(
-                              placeList: widget.placeList,
+                              placeList: widget.placeList!,
                             );
                           },
                         );
