@@ -3,10 +3,19 @@ part of 'saved_lists_bloc.dart';
 abstract class SavedListsState extends Equatable {
   List<PlaceList>? placeLists;
   Stream<List<PlaceList>>? placeListsStream;
-  SavedListsState({this.placeLists});
+  final PlaceList? placeList;
+  final User? listOwner;
+  final List<Place>? places;
+  SavedListsState(
+      {this.placeLists,
+      this.placeListsStream,
+      this.placeList,
+      this.listOwner,
+      this.places});
 
   @override
-  List<Object?> get props => [placeLists, placeListsStream];
+  List<Object?> get props =>
+      [placeLists, placeListsStream, placeList, listOwner, places];
 }
 
 class SavedListsInitial extends SavedListsState {}
