@@ -18,7 +18,6 @@ import 'package:leggo/view/widgets/places/add_place_card.dart';
 import 'package:leggo/view/widgets/places/blank_place_card.dart';
 import 'package:leggo/view/widgets/places/place_card.dart';
 import 'package:leggo/view/widgets/places/search_places_sheet.dart';
-import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:showcaseview/showcaseview.dart';
 
@@ -107,203 +106,120 @@ class _CategoryPageState extends State<CategoryPage> {
                   // buildWhen: (previous, current) =>
                   //     previous.placeList != current.placeList,
                   builder: (context, state) {
-                    if (state is SavedPlacesLoading ||
-                        state is SavedPlacesUpdated) {
-                      rows = [
-                        for (int i = 0; i < 5; i++)
-                          Opacity(
-                            opacity: 0.4,
-                            child: Animate(
-                              onComplete: (controller) {
-                                controller.repeat();
-                              },
-                              effects: const [
-                                //FadeEffect(),
-                                ShimmerEffect(
-                                    duration: Duration(milliseconds: 500),
-                                    curve: Curves.easeInOut),
-                                // SlideEffect(curve: Curves.easeOutBack)
-                              ],
-                              child: const BlankPlaceCard(),
-                            ),
-                          )
-                      ];
+                    // if (state is SavedPlacesLoading ||
+                    //     state is SavedPlacesUpdated) {
+                    //   rows = [
+                    //     for (int i = 0; i < 5; i++)
+                    //       Opacity(
+                    //         opacity: 0.4,
+                    //         child: Animate(
+                    //           onComplete: (controller) {
+                    //             controller.repeat();
+                    //           },
+                    //           effects: const [
+                    //             //FadeEffect(),
+                    //             ShimmerEffect(
+                    //                 duration: Duration(milliseconds: 500),
+                    //                 curve: Curves.easeInOut),
+                    //             // SlideEffect(curve: Curves.easeOutBack)
+                    //           ],
+                    //           child: const BlankPlaceCard(),
+                    //         ),
+                    //       )
+                    //   ];
 
-                      return CustomScrollView(
-                        controller: mainScrollController,
-                        slivers: [
-                          SliverAppBar.medium(
-                            expandedHeight: 125,
-                            // leading: Padding(
-                            //   padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                            //   child: IconButton(
-                            //     onPressed: () {},
-                            //     icon: const Icon(Icons.menu),
-                            //   ),
-                            // ),
-                            title: Row(
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                LoadingAnimationWidget.beat(
-                                    color: Theme.of(context)
-                                        .primaryIconTheme
-                                        .color!,
-                                    size: 20.0),
-                              ],
-                            ),
-                            actions: [
-                              IconButton(
-                                  onPressed: () {},
-                                  icon: const Icon(Icons.more_vert)),
-                            ],
-                          ),
-                          RandomPlaceBar(
-                              keys: [
-                                _visitedFilterShowcase,
-                                _inviteCollaboratorShowcase,
-                                _randomWheelShowcase,
-                                _checklistShowcase
-                              ],
-                              currentPlaceList: null,
-                              controller: controller,
-                              places: null),
-                          SliverList(
-                              delegate: SliverChildListDelegate(
-                            rows
-                                .animate(
-                                  interval: 150.ms,
-                                )
-                                .slideY(
-                                    curve: Curves.easeOutSine,
-                                    begin: -1.0,
-                                    duration: 400.ms)
-                                .fadeIn(
-                                  //delay: 100.ms,
-                                  curve: Curves.easeOutSine,
-                                  //  duration: 600.ms,
-                                ),
-                          ))
-                        ],
-                      );
-                    }
+                    //   return CustomScrollView(
+                    //     controller: mainScrollController,
+                    //     slivers: [
+                    //       SliverAppBar.medium(
+                    //         expandedHeight: 125,
+                    //         // leading: Padding(
+                    //         //   padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                    //         //   child: IconButton(
+                    //         //     onPressed: () {},
+                    //         //     icon: const Icon(Icons.menu),
+                    //         //   ),
+                    //         // ),
+                    //         title: Row(
+                    //           mainAxisSize: MainAxisSize.min,
+                    //           children: [
+                    //             LoadingAnimationWidget.beat(
+                    //                 color: Theme.of(context)
+                    //                     .primaryIconTheme
+                    //                     .color!,
+                    //                 size: 20.0),
+                    //           ],
+                    //         ),
+                    //         actions: [
+                    //           IconButton(
+                    //               onPressed: () {},
+                    //               icon: const Icon(Icons.more_vert)),
+                    //         ],
+                    //       ),
+                    //       RandomPlaceBar(
+                    //           keys: [
+                    //             _visitedFilterShowcase,
+                    //             _inviteCollaboratorShowcase,
+                    //             _randomWheelShowcase,
+                    //             _checklistShowcase
+                    //           ],
+                    //           currentPlaceList: null,
+                    //           controller: controller,
+                    //           places: null),
+                    //       SliverList(
+                    //           delegate: SliverChildListDelegate(
+                    //         rows
+                    //             .animate(
+                    //               interval: 150.ms,
+                    //             )
+                    //             .slideY(
+                    //                 curve: Curves.easeOutSine,
+                    //                 begin: -1.0,
+                    //                 duration: 400.ms)
+                    //             .fadeIn(
+                    //               //delay: 100.ms,
+                    //               curve: Curves.easeOutSine,
+                    //               //  duration: 600.ms,
+                    //             ),
+                    //       ))
+                    //     ],
+                    //   );
+                    // }
                     if (state is SavedPlacesFailed) {
                       return const Center(
                         child: Text('Error Loading List!'),
                       );
                     }
 
-                    if (state is SavedPlacesLoaded) {
-                      User listOwner = state.listOwner;
-                      PlaceList currentPlaceList = state.placeList;
+                    if (state is SavedPlacesLoaded ||
+                        state is SavedPlacesUpdated ||
+                        state is SavedPlacesLoading) {
+                      User? listOwner =
+                          context.watch<SavedListsBloc>().state.listOwner;
+                      PlaceList? currentPlaceList =
+                          context.watch<SavedListsBloc>().state.placeList;
                       List<User>? contributors =
                           context.watch<SavedPlacesBloc>().state.contributors;
                       List<String> contributorAvatars = [];
-                      List<Place> places = state.places;
-
-                      contributorAvatars.add(state.listOwner.profilePicture!);
-                      if (contributors != null) {
-                        for (User user in contributors) {
-                          contributorAvatars.add(user.profilePicture!);
-                        }
-                      }
-
-                      void _onReorder(int oldIndex, int newIndex) {
-                        Place place = state.places.removeAt(oldIndex);
-                        state.places.insert(newIndex, place);
-                        setState(() {
-                          Widget row = rows.removeAt(oldIndex);
-                          rows.insert(newIndex, row);
-                        });
-                      }
-
-                      if (places.isNotEmpty) {
-                        // rows = [
-                        //   for (Place place in places)
-                        //     Padding(
-                        //       padding: const EdgeInsets.symmetric(
-                        //           vertical: 4.0, horizontal: 8.0),
-                        //       child: PlaceCard(
-                        //           place: place,
-                        //           placeList: currentPlaceList,
-                        //           imageUrl: place.mainPhoto,
-                        //           memoryImage: place.mainPhoto,
-                        //           placeName: place.name!,
-                        //           ratingsTotal: place.rating,
-                        //           placeDescription: place.reviews![0]['text'],
-                        //           closingTime: place.hours![0],
-                        //           placeLocation: place.city!),
-                        //     )
-                        // ];
-                        if (state.places.length < 5) {
-                          for (int i = 0; i < 5 - state.places.length; i++) {
-                            rows.add(Opacity(
-                              opacity: 0.4,
-                              child: Animate(effects: const [
-                                ShimmerEffect(
-                                    duration: Duration(milliseconds: 500),
-                                    curve: Curves.easeInOut),
-                              ], child: const BlankPlaceCard()),
-                            ));
+                      List<Place>? places =
+                          context.watch<SavedListsBloc>().state.places;
+                      if (state is SavedPlacesLoaded && listOwner != null) {
+                        contributorAvatars.add(listOwner.profilePicture!);
+                        if (contributors != null) {
+                          for (User user in contributors) {
+                            contributorAvatars.add(user.profilePicture!);
                           }
                         }
-                      } else {
-                        rows = [
-                          for (int i = 0; i < 5; i++)
-                            Opacity(
-                              opacity: 0.4,
-                              child: Animate(effects: const [
-                                ShimmerEffect(
-                                    duration: Duration(milliseconds: 500),
-                                    curve: Curves.easeInOut),
-                                SlideEffect(
-                                  curve: Curves.easeOutBack,
-                                )
-                              ], child: const BlankPlaceCard()),
-                            )
-                        ];
-                        rows.insert(
-                          0,
-                          Padding(
-                            padding: const EdgeInsets.only(top: 8.0),
-                            child: Animate(
-                              effects: const [
-                                SlideEffect(
-                                  curve: Curves.easeOutBack,
-                                ),
-                                ShimmerEffect(
-                                    duration: Duration(milliseconds: 500),
-                                    curve: Curves.easeInOut)
-                              ],
-                              child: const AddPlaceCard(),
-                            ),
-                          ),
-                        );
-                      }
 
-                      return BlocConsumer<EditPlacesBloc, EditPlacesState>(
-                        listener: (context, state) async {
-                          if (state is EditPlacesSubmitted) {
-                            context.read<ListSortCubit>().state.status ==
-                                    'Visited'
-                                ? context.read<SavedPlacesBloc>().add(
-                                    LoadVisitedPlaces(
-                                        placeList: currentPlaceList))
-                                : context.read<SavedPlacesBloc>().add(
-                                    LoadPlaces(placeList: currentPlaceList));
-                            context
-                                .read<SavedListsBloc>()
-                                .add(LoadSavedLists());
-                          }
-                        },
-                        builder: (context, state) {
+                        if (state.places.isNotEmpty) {
                           rows = [
-                            for (Place place in places)
+                            for (Place place in state.places)
                               Padding(
                                 padding: const EdgeInsets.symmetric(
                                     vertical: 4.0, horizontal: 8.0),
                                 child: PlaceCard(
                                     place: place,
-                                    placeList: currentPlaceList,
+                                    placeList: currentPlaceList!,
                                     imageUrl: place.mainPhoto,
                                     memoryImage: place.mainPhoto,
                                     placeName: place.name!,
@@ -313,11 +229,73 @@ class _CategoryPageState extends State<CategoryPage> {
                                     placeLocation: place.city!),
                               )
                           ];
+                          if (state.places.length < 5) {
+                            for (int i = 0; i < 5 - state.places.length; i++) {
+                              rows.add(Opacity(
+                                opacity: 0.4,
+                                child: Animate(effects: const [
+                                  ShimmerEffect(
+                                      duration: Duration(milliseconds: 500),
+                                      curve: Curves.easeInOut),
+                                ], child: const BlankPlaceCard()),
+                              ));
+                            }
+                          }
+                        } else {
+                          rows = [
+                            for (int i = 0; i < 5; i++)
+                              Opacity(
+                                opacity: 0.4,
+                                child: Animate(effects: const [
+                                  ShimmerEffect(
+                                      duration: Duration(milliseconds: 500),
+                                      curve: Curves.easeInOut),
+                                  SlideEffect(
+                                    curve: Curves.easeOutBack,
+                                  )
+                                ], child: const BlankPlaceCard()),
+                              )
+                          ];
+                          rows.insert(
+                            0,
+                            Padding(
+                              padding: const EdgeInsets.only(top: 8.0),
+                              child: Animate(
+                                effects: const [
+                                  SlideEffect(
+                                    curve: Curves.easeOutBack,
+                                  ),
+                                  ShimmerEffect(
+                                      duration: Duration(milliseconds: 500),
+                                      curve: Curves.easeInOut)
+                                ],
+                                child: const AddPlaceCard(),
+                              ),
+                            ),
+                          );
+                        }
+                      }
+
+                      return BlocConsumer<EditPlacesBloc, EditPlacesState>(
+                        listener: (context, state) async {
+                          if (state is EditPlacesSubmitted) {
+                            context.read<ListSortCubit>().state.status ==
+                                    'Visited'
+                                ? context.read<SavedPlacesBloc>().add(
+                                    LoadVisitedPlaces(
+                                        placeList: currentPlaceList!))
+                                : context.read<SavedPlacesBloc>().add(
+                                    LoadPlaces(placeList: currentPlaceList!));
+                            context
+                                .read<SavedListsBloc>()
+                                .add(LoadSavedLists());
+                          }
+                        },
+                        builder: (context, state) {
                           return Stack(
                             alignment: AlignmentDirectional.bottomCenter,
                             children: [
                               CustomScrollView(
-                                cacheExtent: 1000,
                                 controller: mainScrollController,
                                 slivers: [
                                   CategoryPageAppBar(
@@ -335,22 +313,199 @@ class _CategoryPageState extends State<CategoryPage> {
                                       currentPlaceList: currentPlaceList,
                                       controller: controller,
                                       places: places),
-                                  SliverList(
-                                      delegate: SliverChildListDelegate(
-                                    rows
-                                        .animate(
-                                          interval: 150.ms,
+                                  context.watch<SavedPlacesBloc>().state
+                                          is SavedPlacesLoaded
+                                      ? StreamBuilder<List<Place>>(
+                                          stream: context
+                                              .watch<SavedPlacesBloc>()
+                                              .state
+                                              .placeStream,
+                                          builder: (context, snapshot) {
+                                            if (snapshot.hasData == false ||
+                                                snapshot.data == null) {
+                                              return const SliverToBoxAdapter();
+                                            }
+                                            // if (state is SavedPlacesLoading) {
+                                            //   return SliverList(
+                                            //       delegate: SliverChildListDelegate(
+                                            //     [
+                                            //       for (int i = 0; i < 5; i++)
+                                            //         Opacity(
+                                            //           opacity: 0.4,
+                                            //           child: Animate(
+                                            //               effects: const [
+                                            //                 ShimmerEffect(
+                                            //                     duration: Duration(
+                                            //                         milliseconds:
+                                            //                             500),
+                                            //                     curve:
+                                            //                         Curves.easeInOut),
+                                            //                 SlideEffect(
+                                            //                   curve:
+                                            //                       Curves.easeOutBack,
+                                            //                 )
+                                            //               ],
+                                            //               child:
+                                            //                   const BlankPlaceCard()),
+                                            //         )
+                                            //     ],
+                                            //   ));
+                                            // }
+
+                                            if (snapshot.hasData &&
+                                                snapshot.data != null) {
+                                              List<Widget> placeCards = [];
+                                              print(
+                                                  'snapshot has data ${snapshot.data}');
+                                              //rows.clear();
+                                              List<Place> places =
+                                                  snapshot.data!;
+                                              if (places.isNotEmpty &&
+                                                  context
+                                                          .watch<SavedPlacesBloc>()
+                                                          .state
+                                                      is SavedPlacesLoaded) {
+                                                placeCards = [
+                                                  for (Place place in places)
+                                                    Padding(
+                                                      padding: const EdgeInsets
+                                                              .symmetric(
+                                                          vertical: 4.0,
+                                                          horizontal: 8.0),
+                                                      child: PlaceCard(
+                                                          place: place,
+                                                          placeList: context
+                                                              .read<
+                                                                  SavedPlacesBloc>()
+                                                              .state
+                                                              .placeList!,
+                                                          imageUrl:
+                                                              place.mainPhoto,
+                                                          memoryImage:
+                                                              place.mainPhoto,
+                                                          placeName:
+                                                              place.name!,
+                                                          ratingsTotal:
+                                                              place.rating,
+                                                          placeDescription:
+                                                              place.reviews![0]
+                                                                  ['text'],
+                                                          closingTime:
+                                                              place.hours![0],
+                                                          placeLocation:
+                                                              place.city!),
+                                                    )
+                                                ]
+                                                    .animate()
+                                                    .slideY(
+                                                        curve:
+                                                            Curves.easeOutSine,
+                                                        begin: -1.0,
+                                                        duration: 400.ms)
+                                                    .fadeIn(
+                                                      //delay: 100.ms,
+                                                      curve: Curves.easeOutSine,
+                                                      //  duration: 600.ms,
+                                                    );
+                                              }
+
+                                              for (int i = 0; i < 5; i++) {
+                                                rows.add(Opacity(
+                                                  opacity: 0.4,
+                                                  child: Animate(
+                                                      effects: const [
+                                                        ShimmerEffect(
+                                                            duration: Duration(
+                                                                milliseconds:
+                                                                    500),
+                                                            curve: Curves
+                                                                .easeInOut),
+                                                        SlideEffect(
+                                                          curve: Curves
+                                                              .easeOutBack,
+                                                        )
+                                                      ],
+                                                      child:
+                                                          const BlankPlaceCard()),
+                                                ));
+                                              }
+                                              rows.insertAll(0, placeCards);
+                                            } else {
+                                              rows = [
+                                                for (int i = 0; i < 5; i++)
+                                                  Opacity(
+                                                    opacity: 0.4,
+                                                    child: Animate(
+                                                        effects: const [
+                                                          ShimmerEffect(
+                                                              duration: Duration(
+                                                                  milliseconds:
+                                                                      500),
+                                                              curve: Curves
+                                                                  .easeInOut),
+                                                          SlideEffect(
+                                                            curve: Curves
+                                                                .easeOutBack,
+                                                          )
+                                                        ],
+                                                        child:
+                                                            const BlankPlaceCard()),
+                                                  )
+                                              ];
+                                            }
+                                            return SliverList(
+                                                delegate:
+                                                    SliverChildBuilderDelegate(
+                                                        (context, index) =>
+                                                            rows[index],
+                                                        childCount:
+                                                            rows.length));
+
+                                            //     SliverChildListDelegate(
+                                            //   rows
+                                            //       .animate(
+                                            //         interval: 150.ms,
+                                            //       )
+                                            //       .slideY(
+                                            //           curve: Curves.easeOutSine,
+                                            //           begin: -1.0,
+                                            //           duration: 400.ms)
+                                            //       .fadeIn(
+                                            //         //delay: 100.ms,
+                                            //         curve: Curves.easeOutSine,
+                                            //         //  duration: 600.ms,
+                                            //       ),
+                                            // ));
+
+                                            return SliverList(
+                                                delegate:
+                                                    SliverChildListDelegate(
+                                              [
+                                                for (int i = 0; i < 5; i++)
+                                                  Opacity(
+                                                    opacity: 0.4,
+                                                    child: Animate(
+                                                        effects: const [
+                                                          ShimmerEffect(
+                                                              duration: Duration(
+                                                                  milliseconds:
+                                                                      500),
+                                                              curve: Curves
+                                                                  .easeInOut),
+                                                          SlideEffect(
+                                                            curve: Curves
+                                                                .easeOutBack,
+                                                          )
+                                                        ],
+                                                        child:
+                                                            const BlankPlaceCard()),
+                                                  )
+                                              ],
+                                            ));
+                                          },
                                         )
-                                        .slideY(
-                                            curve: Curves.easeOutSine,
-                                            begin: -1.0,
-                                            duration: 400.ms)
-                                        .fadeIn(
-                                          //delay: 100.ms,
-                                          curve: Curves.easeOutSine,
-                                          //  duration: 600.ms,
-                                        ),
-                                  ))
+                                      : const SliverToBoxAdapter(
+                                          child: SizedBox()),
                                 ],
                               ),
                               if (state is EditPlacesStarted ||
@@ -393,13 +548,13 @@ class _CategoryPageState extends State<CategoryPage> {
                                                                   EditPlacesBloc>()
                                                               .add(DeleteVisitedPlaces(
                                                                   placeList:
-                                                                      currentPlaceList))
+                                                                      currentPlaceList!))
                                                           : context
                                                               .read<
                                                                   EditPlacesBloc>()
                                                               .add(DeletePlaces(
                                                                   placeList:
-                                                                      currentPlaceList));
+                                                                      currentPlaceList!));
                                                       // context.read<EditPlacesBloc>().add(
                                                       //     FinishEditing(places: places, placeList: placeList));
                                                     },
@@ -431,7 +586,7 @@ class _CategoryPageState extends State<CategoryPage> {
                                                               EditPlacesBloc>()
                                                           .add(MarkVisitedPlaces(
                                                               placeList:
-                                                                  currentPlaceList));
+                                                                  currentPlaceList!));
                                                       // context.read<EditPlacesBloc>().add(
                                                       //     FinishEditing(places: places, placeList: placeList));
                                                     },
